chore(root): improves file extension readability
