+++ src/config/validation/types.ts
@@ -1,0 +1,6 @@
+export type FieldValidator = (value: unknown) => boolean;
+
+export type FieldValidation = {
+  field: string;
+  validator: FieldValidator;
+};
