+++ src/core/security/providers/index.ts
@@ -3,4 +3,3 @@ export { GitHubSecurityProvider } from "./github";
 export { SnykCLIProvider } from "./snyk";
 export { SocketCLIProvider } from "./socket";
 export { SpektionProvider } from "./spektion";
-export { PackageManagerAuditProvider } from "./package-manager-audit";
