+++ tests/unit/core/security/providers/package-manager-audit.test.ts
@@ -1,7 +1,7 @@
 import { assertMatches, errorIncludes, mock, objectContaining, spyOn } from "../../../setup";
 import { test, describe, afterEach } from "node:test";
 import assert from "node:assert/strict";
-import { PackageManagerAuditProvider } from "../../../../../src/core/security/providers/package-manager-audit";
+import { PackageManagerAuditProvider } from "../../../../../src/providers";
 import type { NpmAuditResult, YarnAuditLine } from "../../../../../src/types";
 
 afterEach(() => {
